Add files via upload
Question 1 [35 marks]

Part 1: [5 marks]

Declare a function with two input parameters: "a" and "b" both integer values. The function returns a random integer in the range [a, b].

Part 2: [20 marks]

Use the function you defined in Part 1 to write a program for the "Guess the Number" game. In this game, the user is to guess a random number in the range [1, 100] generated by the computer. Here are the steps that the program takes:

The program calls the function declared in Part 1 to generate a random integer in the range [1, 100].
The program then asks the user to guess the number generated in the previous step. 
If the user enters the correct number, the program alerts "You won!" and terminates.
If the number entered by the user is not in the range [1, 100], the program alerts an error message and goes to Step 2.
If the number entered by the user is less than the random number generated in Step 1, the program displays "Enter a larger value" and goes to Step 2.
If the number entered by the user is greater than the random number generated in Step 1, the program displays "Enter a smaller value" and goes to Step 2.
The program continues until the user guesses the number correctly.

Part 3: [10 marks]

Rewrite the program written in Part 2 to make the program stop after 20 wrong guesses. If the user enters 20 wrong guesses, the program alerts the message "You lost!" and terminates.

Question 2 [15 marks]

Write a function expression named "isPrime" that returns "true" if an input integer greater than 1 is prime, and returns "false", otherwise. 

A prime integer is an integer with no division other than 1 and itself. For example, 7, 13, 11, and 2 are all prime numbers, while 6, 21, and 44 are not prime.

To identify if "n" is a prime number or not, repeatedly divide it by the integers in the range of [2,n]. If it is not divisible by any of these numbers, it is prime. If at least one of these values divides "n" evenly, "n" is not prime.