Add files via upload
Instructions
Declare a function with three input parameters: "n" which is a number, "ch" which is a character, and "choice" which is "r", "c", "fs", or "hs". The function proceeds as follows:

If "choice" is "r", the function logs a row including "n" occurrences of "ch". For example, if n = 5 and ch = "*" the output is:
*****

If "choice" is "c", the function logs a column including "n" occurrences of "ch". For example, if n = 5 and ch = "*" the output is:
*

*

*

*

*

If "choice" is "fs", the function logs a filled square of the size "n" of "ch". For example, if n = 5 and ch = "*" the output is:
*****

*****

*****

*****

*****

If "choice" is "hs", the function logs a hollow square of the size "n" of "ch". For example, if n = 5 and ch = "*" the output is:
*****

*   *

*   *

*   *

*****

If any of the input parameters are invalid, the function logs a detailed error message showing all the invalid values. For example, if n = -2, ch = "O", and choice = "xxx" the output is:
Error: '-2' is not valid for the number of characters. 'xxx' is not a valid choice.

Program requirements:

By default, the function displays a row including a single "?" character.
At the start of the function, define 4 function expressions that return a string: a function to return a row of n characters, a function to return a column of characters, a function to return a filled square of characters and a function to return a hollow square of characters. Then use these function expressions in the rest of your code to complete the main function.
If the argument passed for "n" is not a whole number, use its floor. For example, if n = 6.7, use the value of 6 for "n".