Add files via upload
Q1: DOM [40 marks]

Create a web page that first asks the user to enter a positive integer as "n" and displays an nxn times table on the web page. This program requires you to use DOM to create a table that has n + 1 rows and n + 1 columns. The top row and the left-most column are headings labeled by numbers from 1 to n. The rest of the table includes the times table's values. A sample output for the value of n = 8 is shown below. If the user enters an invalid value for "n", the program keeps asking for a valid value (i.e. n >= 1).



Q2: Canvas [30 marks]

Create a web page including a canvas. Draw your favorite carved pumpkin in the center of your canvas. Color your pumpkin in orange with black carvings. More complex carvings receive a higher mark. Feel free to be innovative in creating a spooky scene.